Various bug fixes

+++ debugger.cpp
@@ -6,10 +6,10 @@
 //WARNING: VERRRY SLOW
 std::ostream& operator<<(std::ostream& os, CPUDebugger& cpu) {
 	return os << "Regs:\n" << 
-	"\tPC: " << std::hex << std::setw(4) << std::setfill('0') << cpu.getPC() <<
-	"\tA:  " << std::hex << std::setw(4) << std::setfill('0') << cpu.getA() << 
-	"\tX:  " << std::hex << std::setw(4) <<  std::setfill('0') << cpu.getX() <<
-	"\tY:  " << std::hex << std::setw(4) <<  std::setfill('0') << cpu.getY() <<
+	"\tPC: " << std::hex << std::setw(4) << std::setfill('0') << (int) cpu.getPC() <<
+	"\tA:  " << std::hex << std::setw(4) << std::setfill('0') << (int) cpu.getA() << 
+	"\tX:  " << std::hex << std::setw(4) <<  std::setfill('0') << (int) cpu.getX() <<
+	"\tY:  " << std::hex << std::setw(4) <<  std::setfill('0') << (int) cpu.getY() <<
 	"\n\nInstruction: " << std::hex << std::setw(4) << 
 	cpu.getInstr() << std::endl;
 }
